+++ cli/src/installers/prisma.ts
@@ -1,6 +1,7 @@
-import fs from "fs-extra";
 import path from "path";
+import fs from "fs-extra";
 import { type PackageJson } from "type-fest";
+
 import { PKG_ROOT } from "~/consts.js";
 import { type Installer } from "~/installers/index.js";
 import { addPackageDependency } from "~/utils/addPackageDependency.js";
@@ -26,16 +27,18 @@ export const prismaInstaller: Installer = ({ projectDir, packages }) => {
   );
   const schemaDest = path.join(projectDir, "prisma/schema.prisma");
 
-  const clientSrc = path.join(extrasDir, "src/server/db.ts");
+  const clientSrc = path.join(extrasDir, "src/server/db/db-prisma.ts");
   const clientDest = path.join(projectDir, "src/server/db.ts");
 
-  // add postinstall script to package.json
+  // add postinstall and push script to package.json
   const packageJsonPath = path.join(projectDir, "package.json");
 
   const packageJsonContent = fs.readJSONSync(packageJsonPath) as PackageJson;
   packageJsonContent.scripts = {
     ...packageJsonContent.scripts,
     postinstall: "prisma generate",
+    "db:push": "prisma db push",
+    "db:studio": "prisma studio",
   };
 
   fs.copySync(schemaSrc, schemaDest);