+++ upgrade/prettier.config.mjs
@@ -1,0 +1,18 @@
+import baseConfig from "../prettier.config.mjs";
+
+/**
+ * @type {import('prettier').Config & import('prettier-plugin-tailwindcss').options &
+ *       import("@ianvs/prettier-plugin-sort-imports").PluginConfig}
+ */
+const config = {
+  ...baseConfig,
+  plugins: [
+    "@ianvs/prettier-plugin-sort-imports",
+    "prettier-plugin-tailwindcss",
+  ],
+  importOrder: ["<THIRD_PARTY_MODULES>", "", "^~/", "^[.][.]/", "^[.]/"],
+  importOrderParserPlugins: ["typescript", "jsx", "decorators-legacy"],
+  importOrderTypeScriptVersion: "4.4.0",
+};
+
+export default config;