+++ www/src/components/navigation/LanguageSelect.tsx
@@ -1,8 +1,9 @@
-import { type KnownLanguageCode, KNOWN_LANGUAGES } from "../../config";
-import { getIsRtlFromLangCode } from "../../languages";
 import { Listbox, Transition } from "@headlessui/react";
 import clsx from "clsx";
 import { Fragment } from "react";
+
+import { KNOWN_LANGUAGES, type KnownLanguageCode } from "../../config";
+import { getIsRtlFromLangCode } from "../../languages";
 
 interface LanguageSelectProps {
   language: KnownLanguageCode;